+++ assets/Mods/ETBCommandsMod/Scripts/aliases.lua
@@ -22,7 +22,7 @@ local function ProcessHelp(FullCommand, Parameters, Ar)
 
     InitMod()
 
-    Log([[All custom commands (Command : Alias), case-insensetive:
+    Log([[All custom commands (Command : Alias), case-sensitive:
         give : g (Give an item directly into the inventory)
         drop : d (Drop an item on the ground)
         juice : j (Toggle infinite juice effect)
@@ -30,11 +30,14 @@ local function ProcessHelp(FullCommand, Parameters, Ar)
         toggledebugcamera : dc (Toggle debug camera)
         teleport : t (Teleport to where you're looking)
         fullbright : fb (Toggle fullbright)
-        noclip (Toggle noclip mode)
+        noclip : nc (Toggle noclip mode)
+        getpos (Copy current coordinates)
+        setpos (Teleport to coordinates)
         bind (Bind a command to a key)
         binds (Print all current binds)
         unbind (Remove a custom bind)
         unbindall (Remove all custom binds)
+        listkeys (List of all possible keys for bind)
         help : aliases (This message)]])
     return true
 end